# openapi.yaml
...
openapi: 3.0.3
info:
    title: Go-Kratos Console API
    description: Management console for Go-Kratos
    contact:
        name: Sendya
        url: https://github.com/sendya
        email: yladmxa@gmail.com
    license:
        name: Apache License
        url: https://github.com/google/gnostic/blob/master/LICENSE
    version: 1.0.0
paths:
    /api/console/app:
        get:
            tags:
                - App
            operationId: App_List
            parameters:
                - name: pagination.current
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.page_size
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.total
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pagination.raw_total
                  in: query
                  schema:
                    type: string
                - name: namespace_id
                  in: query
                  description: 命名空间ID
                  schema:
                    type: string
                - name: name
                  in: query
                  description: 应用名称
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.AppListReply'
        post:
            tags:
                - App
            operationId: App_Create
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.AppInfo'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.AppCreateReply'
    /api/console/app/-/metas:
        get:
            tags:
                - App
            operationId: App_NamespaceAppList
            parameters:
                - name: name
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.NamespaceAppListReply'
    /api/console/app/{id}:
        get:
            tags:
                - App
            operationId: App_Get
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.AppGetReply'
        put:
            tags:
                - App
            operationId: App_Update
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.console.resource.AppUpdateRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.AppUpdateReply'
        delete:
            tags:
                - App
            operationId: App_Delete
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.console.resource.AppDeleteReply'
...
components:
    schemas:
...
        api.console.resource.AppCreateReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.AppInfo'
        api.console.resource.AppDeleteReply:
            type: object
            properties: {}
        api.console.resource.AppGetReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.AppInfo'
        api.console.resource.AppInfo:
            type: object
            properties:
                id:
                    type: string
                    description: 唯一键
                name:
                    type: string
                    description: 应用名称
                alias:
                    type: string
                    description: 应用别名
                description:
                    type: string
                    description: 应用描述
                icon:
                    type: string
                    description: 应用图标
                ports:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.console.resource.AppInfo_AppPort'
                    description: 应用端口
                type:
                    type: integer
                    description: 应用类型
                    format: enum
                namespace_id:
                    type: string
                    description: 命名空间id
                users:
                    type: array
                    items:
                        type: string
                    description: 应用负责人
                repos:
                    type: array
                    items:
                        type: string
                    description: 应用仓库
                created_at:
                    type: string
                    description: 创建时间
                updated_at:
                    type: string
                    description: 修改时间
        api.console.resource.AppInfo_AppPort:
            type: object
            properties:
                port:
                    type: integer
                    description: 端口 0代表随机端口
                    format: uint32
                protocol:
                    type: integer
                    description: 协议
                    format: enum
                remark:
                    type: string
                    description: 备注
            description: 应用端口描述
        api.console.resource.AppListReply:
            type: object
            properties:
                pagination:
                    $ref: '#/components/schemas/protobuf.Pagination'
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.console.resource.AppInfo'
        api.console.resource.AppUpdateReply:
            type: object
            properties:
                data:
                    $ref: '#/components/schemas/api.console.resource.AppInfo'
        api.console.resource.AppUpdateRequest:
            type: object
            properties:
                id:
                    type: string
...
        api.console.resource.NamespaceApp:
            type: object
            properties:
                id:
                    type: string
                name:
                    type: string
                alias:
                    type: string
                type:
                    type: integer
                    format: enum
                namespace_id:
                    type: string
                namespace_name:
                    type: string
                namespace_alias:
                    type: string
        api.console.resource.NamespaceAppListReply:
            type: object
            properties:
                data:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.console.resource.NamespaceApp'
...
        protobuf.Pagination:
            type: object
            properties:
                current:
                    type: integer
                    format: int32
                page_size:
                    type: integer
                    format: int32
                total:
                    type: integer
                    format: int32
                raw_total:
                    type: string
            description: 通用的分页信息
    securitySchemes:
        ApiKeyAuth:
            type: apiKey
            name: Authorization
            in: header
...